backend/requirements.txt
# Core
fastapi==0.121.1
uvicorn[standard]==0.38.0
python-multipart==0.0.20
pydantic==2.12.4
pydantic-settings==2.12.0
python-dotenv==1.2.1

# Database & ORM
sqlmodel==0.0.27
sqlalchemy==2.0.44
asyncpg==0.30.0
alembic==1.17.1
psycopg2-binary==2.9.11
pymongo==4.15.4


# Supabase
supabase==2.24.0
postgrest==2.24.0
gotrue==2.12.4
python-jose==3.5.0

# Security & Auth
email-validator==2.2.0

# Data Processing
pandas==2.3.3
numpy==2.3.4

# File Handling
aiofiles==25.1.0
python-magic==0.4.27

# Vector/AI
sentence-transformers==5.1.2
openai==2.7.2
# onnxruntime==1.23.2
# chromadb==1.3.4

# Testing
pytest==9.0.1
pytest-asyncio==1.3.0
httpx==0.28.1